Remove more v2 directories in cypress shared
