+++ cypress_shared/pages/temporary-accommodation/manage/premisesArchive.ts
@@ -1,5 +1,5 @@
-import Page from '../../../page'
-import { Cas3Premises } from '../../../../../server/@types/shared'
+import Page from '../../page'
+import { Cas3Premises } from '../../../../server/@types/shared'
 
 export default class PremisesArchivePage extends Page {
   constructor(premises: Cas3Premises) {
