+++ cypress_shared/pages/temporary-accommodation/manage/premisesUnarchive.ts
@@ -1,5 +1,5 @@
-import Page from '../../../page'
-import { Cas3Premises } from '../../../../../server/@types/shared'
+import Page from '../../page'
+import { Cas3Premises } from '../../../../server/@types/shared'
 
 export default class PremisesUnarchivePage extends Page {
   constructor(premises: Cas3Premises) {
