--- cypress_shared/pages/temporary-accommodation/manage/v2/premisesNew.ts
@@ -1,13 +1,0 @@
-import paths from '../../../../../server/paths/temporary-accommodation/manage'
-import PremisesEditablePage from './premisesEditable'
-
-export default class PremisesNewPage extends PremisesEditablePage {
-  constructor() {
-    super('Add a property')
-  }
-
-  static visit(): PremisesNewPage {
-    cy.visit(paths.premises.new({}))
-    return new PremisesNewPage()
-  }
-}
